Initial real commit
I mostly completed all of this in an hour or two, it was relatively simple. This tool is not meant to be much more than a simple Qt UI to PNG file generator. This will produce a .png file with a given Qt .ui widget or window painted in it, at any given size and quality. It seems to work well enough with basic testing, but I haven't done anything thorough with this and frankly it's probably broken in several ways. Don't rely on this tool!

There is some basic preparation for translation, but no translations exist.